requirements.txt
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.3.0
Flask==1.0.2
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
MarkupSafe==1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
python-decouple==3.1
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
